Parse random styled-components classNames into respective constant names
