+++ package.json
@@ -32,6 +32,7 @@
     "webpack-cli": "^3.3.11"
   },
   "devDependencies": {
+    "babel-plugin-styled-components": "^1.10.7",
     "eslint": "^7.3.0",
     "eslint-config-airbnb": "^18.2.0",
     "eslint-plugin-import": "^2.21.2",
